fix: integer columns in articles and queue workers

+++ database/migrations/2020_11_24_195823_create_articles_table.php
@@ -16,9 +16,9 @@ class CreateArticlesTable extends Migration
             $table->string('path');
             $table->string('url');
             $table->string('canonical_url');
-            $table->unsignedInteger('comments_count');
-            $table->unsignedInteger('public_reactions_count');
-            $table->unsignedInteger('positive_reactions_count');
+            $table->unsignedBigInteger('comments_count');
+            $table->unsignedBigInteger('public_reactions_count');
+            $table->unsignedBigInteger('positive_reactions_count');
             $table->string('tags');
             $table->longText('body_html');
             $table->timestamp('published_timestamp');
